# Improve performance of MOI.copy_to