Part of https://github.com/jump-dev/JuMP.jl/issues/2817

If we include https://github.com/jump-dev/MathOptInterface.jl/pull/1704

Script
```Julia
using MathOptInterface
const MOI = MathOptInterface
using GLPK

function create_model(I, T = 10000)
    model = MOI.Utilities.Model{Float64}()
    v = MOI.VariableIndex[]
    for _ in 1:I
        x = MOI.add_variables(model, T)
        MOI.add_constraint.(model, x, MOI.GreaterThan(0.0))
        MOI.add_constraint.(model, x, MOI.LessThan(100.0))
        for t in 2:T
            f = MOI.ScalarAffineFunction(
                MOI.ScalarAffineTerm.([1.0, -1.0], [x[t], x[t-1]]),
                0.0,
            )
            MOI.add_constraint(model, f, MOI.GreaterThan(-10.0))
            MOI.add_constraint(model, f, MOI.LessThan(10.0))
        end
        append!(v, x)
    end
    g = MOI.ScalarAffineFunction(MOI.ScalarAffineTerm.(1.0, v), 0.0)
    MOI.set(model, MOI.ObjectiveSense(), MOI.MIN_SENSE)
    MOI.set(model, MOI.ObjectiveFunction{typeof(g)}(), g)
    dest = GLPK.Optimizer()
    MOI.copy_to(dest, model)
    return dest
end
```

Before
```Julia
julia> @benchmark create_model(100, 1000)
BenchmarkTools.Trial: 11 samples with 1 evaluation.
 Range (min … max):  348.692 ms … 690.298 ms  ┊ GC (min … max): 23.52% … 55.39%
 Time  (median):     468.977 ms               ┊ GC (median):    34.47%
 Time  (mean ± σ):   510.078 ms ± 141.469 ms  ┊ GC (mean ± σ):  44.43% ± 15.48%

  ▁  ▁ █   ▁           ▁                           █    ▁  ▁  ▁  
  █▁▁█▁█▁▁▁█▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁█▁▁█▁▁█ ▁
  349 ms           Histogram: frequency by time          690 ms <

 Memory estimate: 384.43 MiB, allocs estimate: 2898897.
```

After
```Julia
julia> @benchmark create_model(100, 1000)
BenchmarkTools.Trial: 20 samples with 1 evaluation.
 Range (min … max):  184.094 ms … 390.970 ms  ┊ GC (min … max): 12.39% … 19.73%
 Time  (median):     240.240 ms               ┊ GC (median):    24.33%
 Time  (mean ± σ):   253.573 ms ±  43.221 ms  ┊ GC (mean ± σ):  24.21% ±  5.05%

                ▃██       ▃  ▃                                   
  ▇▁▁▁▁▁▁▁▁▁▇▁▇▇███▁▁▇▇▁▁▁█▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▇ ▁
  184 ms           Histogram: frequency by time          391 ms <

 Memory estimate: 152.62 MiB, allocs estimate: 1299738.
```